fix: add test for getCurrentUser
