+++ Supertab.test.ts
@@ -1,6 +1,7 @@
 import { test, expect, describe } from "bun:test";
 import { rest, server } from "@/mocks/server";
 import Supertab from ".";
+import { UserResponse, UserResponseToJSON } from "@laterpay/tapper-sdk";
 
 describe("Supertab", () => {
   describe("SupertabInit", () => {
@@ -25,6 +26,39 @@ describe("Supertab", () => {
         ),
       );
       expect(await client.getApiVersion()).toBe("1.0.0");
+    });
+  });
+
+  describe(".getCurrentUser", () => {
+    test("return logged user from tapper", async () => {
+      localStorage.setItem(
+        "supertab-auth",
+        JSON.stringify({
+          expiresAt: Date.now() + 100000,
+        }),
+      );
+
+      const client = new Supertab({ clientId: "test-client-id" });
+
+      const user: UserResponse = {
+        id: "test-user-id",
+        firstName: "Test",
+        lastName: "User",
+        createdAt: new Date("2021-01-01T00:00:00.000Z"),
+        updatedAt: new Date("2021-01-01T00:00:00.000Z"),
+        deletedAt: new Date("2021-01-01T00:00:00.000Z"),
+        active: true,
+        isEmailVerified: true,
+      };
+
+      server.use(
+        rest.get(
+          "https://tapi.sbx.laterpay.net/v1/identity/me",
+          (_, res, ctx) =>
+            res(ctx.status(200), ctx.json(UserResponseToJSON(user))),
+        ),
+      );
+      expect(await client.getCurrentUser()).toEqual(user);
     });
   });
 });
