+++ src/__tests__/auth.test.ts
@@ -184,16 +184,6 @@ describe("auth", () => {
   });
 
   describe("getAuthStatus", () => {
-    beforeEach(() => {
-      const map = new Map();
-      const localStorage = {
-        getItem: map.get.bind(map),
-        setItem: map.set.bind(map),
-        removeItem: map.delete.bind(map),
-      };
-      global.localStorage = localStorage as unknown as Storage;
-    });
-
     it("returns 'valid' if access token is valid", async () => {
       localStorage.setItem(
         "supertab-auth",
