+++ src/setupTests.ts
@@ -1,4 +1,4 @@
-import { beforeAll, afterEach, afterAll } from "bun:test";
+import { beforeAll, beforeEach, afterEach, afterAll } from "bun:test";
 import { server } from "./mocks/server";
 import { GlobalWindow } from "happy-dom";
 
@@ -7,6 +7,16 @@ global.window = new GlobalWindow() as any;
 beforeAll(() => {
   // Enable the mocking in tests.
   server.listen();
+});
+
+beforeEach(() => {
+  const map = new Map();
+  const localStorage = {
+    getItem: map.get.bind(map),
+    setItem: map.set.bind(map),
+    removeItem: map.delete.bind(map),
+  };
+  global.localStorage = localStorage as unknown as Storage;
 });
 
 afterEach(() => {
